reogranizes: removes extra page rendering main birthdayfriend page, extras item into component
